Move files to a new directory under Stub and add register_fprime_implementation for StandardAtomic
