+++ Os/Stub/atomic/atomic.hpp
@@ -6,7 +6,7 @@
 #define OS_STUB_ATOMIC_HPP
 
 namespace std {
-    #include "Os/Stub/atomic_impl.hpp"
+    #include "Os/Stub/atomic/atomic_impl.hpp"
 } // namespace std
 
 #endif // OS_STUB_ATOMIC_HPP
