+++ Os/Stub/atomic/atomic_impl.hpp
